Init with work from January setup
Adds all the files that are currently in my dotfiles that have been
working since January setup of new mac. Not all of it is complete
however I do not have the time or focus to complete. Will carry on work
the next time I need to configure these files.